requirements.txt
aioredis==2.0.1
asgiref==3.5.2
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
black==22.10.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
click==8.1.3
constantly==15.1.0
cryptography==38.0.1
daphne==4.0.0
decorator==5.1.1
Deprecated==1.2.13
dj-database-url==1.0.0
Django==4.1.2
django-cors-headers==3.13.0
django-elasticsearch-dsl==7.3
django-elasticsearch-dsl-drf==0.22.5
django-extensions==3.2.1
django-filter==22.1
django-nine==0.2.7
django-redis==5.2.0
django-stubs==1.14.0
django-stubs-ext==0.7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
elastic-transport==8.4.0
elasticsearch==7.17.9
elasticsearch-dsl==7.4.0
executing==1.1.1
Faker==17.0.0
flake8==5.0.4
GDAL==3.0.4
geographiclib==2.0
geopy==2.3.0
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.0.0
incremental==22.10.0
ipdb==0.13.9
ipython==8.5.0
isort==5.10.1
jedi==0.18.1
Markdown==3.4.1
matplotlib-inline==0.1.6
mccabe==0.7.0
msgpack==1.0.4
mypy==0.982
mypy-extensions==0.4.3
numpy==1.17.4
packaging==21.3
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.31
psycopg2==2.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
Pygments==2.13.0
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
redis==4.5.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.5.1
toml==0.10.2
tomli==2.0.1
traitlets==5.5.0
Twisted==22.8.0
txaio==22.2.1
types-pytz==2022.5.0.0
types-PyYAML==6.0.12
typing-extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.5
websocket==0.2.1
websocket-client==1.5.1
whitenoise==6.2.0
wrapt==1.14.1
zipp==3.9.0
zope.event==4.6
zope.interface==5.5.0
